time tracking for task item implemented, huge refactoring into components from index into own files
