+++ Client/Shared/TimerComponent.razor
@@ -1,0 +1,46 @@
+﻿@using Recurop
+
+<div style="text-align:center;">
+    <h1>
+        <strong>Flowmodoro</strong>
+    </h1>
+
+    <p>
+        Work time is devided by 5<br />
+        Stop and Swap Mode to take a break <br />
+        If you dont want to take a break just stop and reset the timer :)
+    </p>
+
+</div>
+<div style="margin-top:4em; text-align:center;">
+    <h2>@TitleText</h2>
+    <p style="white-space: pre-wrap">@(string.IsNullOrEmpty(SubtitleText) ? " " : SubtitleText)</p>
+    
+    <div >
+        <p style="visibility: @(_timerVisible ? "visible" : "hidden"); font-size: 75px;">@this._displayTime</p>
+        <div>
+            <button class="btn btn-sm btn-primary" disabled=@(!StopWatchOperation.CanBeStarted || !TimerOpertation.CanBeStarted ? true : false)
+            @onclick="StartTimer">
+                Start
+            </button>
+            <button class="btn btn-sm btn-danger" disabled=@(StopWatchOperation.IsRecurring || TimerOpertation.IsRecurring ? false : true)
+            @onclick="StopTimer">
+                Stop
+            </button>
+            <button class="btn btn-sm btn-danger" disabled=@(StopWatchOperation.IsRecurring || TimerOpertation.IsRecurring ? true : false)
+            @onclick="ResetTimer">
+                Reset
+            </button>
+            <button class="btn btn-sm btn_danger" disabled=@(!CanBeSwapped())
+            @onclick="SwapMode">
+                SwapMode
+            </button>
+        </div>
+    </div>
+    @if (BlinkerOperation?.Status == RecurringOperationStatus.Executing)
+    {
+        <div>
+            <audio autoplay loop><source src="/Assets/Audio/30-seconds-alarm.mp3" /></audio>
+        </div>
+    }
+</div>
